test IQU cube maker

+++ code/make_rht_cube.py
@@ -170,10 +170,10 @@ class Cube():
         priheader_U = hdulist_U[0].header
         
         # start and end velocities in meters
-        startvel = galfa_vel_helpers.get_galfa_W_truevel(974)*1000
-        stopvel = galfa_vel_helpers.get_galfa_W_truevel(1078)*1000
+        startvel = np.float(galfa_vel_helpers.get_galfa_W_truevel(974))*1000
+        stopvel = np.float(galfa_vel_helpers.get_galfa_W_truevel(1078))*1000
         print(startvel, stopvel)
-        velstep = (np.float(stopvel) - np.float(startvel))/self.nchannels
+        velstep = (stopvel - startvel)/self.nchannels
         
         for priheader in [priheader_I, priheader_Q, priheader_U]:
